@@ -1,15 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>net5.0-windows;net6.0-windows;net7.0-windows</TargetFrameworks>
-		<UseWPF>true</UseWPF>
+		<TargetFrameworks>net8.0</TargetFrameworks>
 		<Platforms>AnyCPU;x64;x86</Platforms>
 		<Configurations>Debug;Release;ReleaseDebug;DebugRelease</Configurations>
 		<LangVersion>default</LangVersion>
 	</PropertyGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\HLab.Options\HLab.Options.csproj" />
+	  <ProjectReference Include="..\..\HLab.Core\HLab.Options\HLab.Options.csproj" />
 	</ItemGroup>
 
 </Project>
